==> Notebooks/02 Data Analysis/02_Regression/19_LassoRegression.ipynb <==
# Data Analysis (Regression): Lasso Regression
Lasso Regression, or Least Absolute Shrinkage and Selection Operator, is a modified version of Linear Regression where the sum of the absolute values of the model coefficients is limited in the loss function to minimze the complexity of the model.

## Import Libraries

# Import Required Modules and Packages
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib
import sys

from sklearn import metrics
from sklearn import linear_model
from sklearn.metrics import r2_score
from sklearn.linear_model import Lasso

cd

## Load Data

# Upload the X and y Data
X_train = pd.read_csv('Project/Cleaned/Regression/X_train.csv', na_values = ['..'])
X_test = pd.read_csv('Project/Cleaned/Regression/X_test.csv', na_values = ['..'])
y_train = pd.read_csv('Project/Cleaned/Regression/y_train.csv', na_values = ['..'])
y_test = pd.read_csv('Project/Cleaned/Regression/y_test.csv', na_values = ['..'])

# Drop Column
X_train.drop(['Unnamed: 0'], axis=1, inplace=True)
X_test.drop(['Unnamed: 0'], axis=1, inplace=True)
y_train.drop(['Unnamed: 0'], axis=1, inplace=True)
y_test.drop(['Unnamed: 0'], axis=1, inplace=True)

# Load Dictionary
dictionary_series = np.load('Project/Cleaned/dictionary_series.npy',allow_pickle='TRUE').item()

# Shape of Train and Test Data
print('X_test:', X_test.shape)
print('y_test:', y_test.shape)
print('X_train:', X_train.shape)
print('y_train:', y_train.shape)

## Lasso Regression

# Create and Fit Model
model_lasso = Lasso(alpha=0.01).fit(X_train, y_train.values.ravel()) 

# Model Coefficients
model_coefficients_lasso = model_lasso.coef_

# Model Prediction
y_prediction_train = model_lasso.predict(X_train)
y_prediction_test = model_lasso.predict(X_test)

## Evaluate Model

# Model Results
model_results = [['R-Squared', r2_score(y_train, y_prediction_train), r2_score(y_test, y_prediction_test)],
                 ['Mean Absolute Error', metrics.mean_absolute_error(y_train, y_prediction_train), metrics.mean_absolute_error(y_test, y_prediction_test)],
                 ['Mean Squared Error', metrics.mean_squared_error(y_train, y_prediction_train), metrics.mean_squared_error(y_test, y_prediction_test)],
                 ['Root Mean Squared Error', np.sqrt(metrics.mean_squared_error(y_train, y_prediction_train)), np.sqrt(metrics.mean_squared_error(y_test, y_prediction_test))]]
model_results_lasso = pd.DataFrame(model_results, columns=['Evaluation Metric', 'Train', 'Test'])
model_results_lasso

# Model Coefficients
model_coefficients_lasso_table = pd.DataFrame(model_lasso.coef_.flatten(), X_train.columns, columns=['Coefficient'])
model_coefficients_lasso_table

# Series Dictionary
dictionary_series['DT.NFL.UNDP.CD']
