==> Notebooks/02 Data Analysis/02_Regression/23_Stacking.ipynb <==
# Data Analysis (Regression): Ensemble Methods - Stacking

## Import Libraries

# Import Required Modules and Packages
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib
import sys

from math import sqrt
from sklearn import metrics
from sklearn.metrics import mean_squared_error 
from sklearn.ensemble import StackingRegressor, RandomForestRegressor
from sklearn.svm import LinearSVR
from sklearn.linear_model import LinearRegression, LassoCV, RidgeCV, ElasticNetCV

random = 20200910

cd

## Load Data

# Upload the X and y Data
X_train = pd.read_csv('Project/Cleaned/Regression/X_train.csv', na_values = ['..'])
X_test = pd.read_csv('Project/Cleaned/Regression/X_test.csv', na_values = ['..'])
y_train = pd.read_csv('Project/Cleaned/Regression/y_train.csv', na_values = ['..'])
y_test = pd.read_csv('Project/Cleaned/Regression/y_test.csv', na_values = ['..'])

# Drop Column
X_train.drop(['Unnamed: 0'], axis=1, inplace=True)
X_test.drop(['Unnamed: 0'], axis=1, inplace=True)
y_train.drop(['Unnamed: 0'], axis=1, inplace=True)
y_test.drop(['Unnamed: 0'], axis=1, inplace=True)

# Load Dictionary
dictionary_series = np.load('Project/Cleaned/dictionary_series.npy',allow_pickle='TRUE').item()

# Shape of Train and Test Data
print('X_test:', X_test.shape)
print('y_test:', y_test.shape)
print('X_train:', X_train.shape)
print('y_train:', y_train.shape)

## Stacking

# Create Model and Fit Model
estimators = [
    ('svr', LinearSVR(random_state=random)),
    ('lasso', LassoCV(random_state=random)),
    ('ridge', RidgeCV()),
    ('e_net', ElasticNetCV(random_state=random))
]

model_sta = StackingRegressor(estimators=estimators, final_estimator=RandomForestRegressor(n_estimators=10, random_state=random)).fit(X_train, y_train.values.ravel())

# Model Predictions
model_predictions_sta = model_sta.predict(X_train)

## Evaluate Model

# Model Metrics
model_rmse_sta = sqrt(mean_squared_error(y_test, model_sta.predict(X_test)))
model_errors_sta = abs(model_predictions_sta.flatten()-y_train.values.ravel())
model_mae_sta = round(np.mean(model_errors_sta), 2)
model_mape_sta = 100 * (model_errors_sta / y_train.values.ravel())
model_accuracy_sta = 100 - np.mean(model_mape_sta[np.isfinite(model_mape_sta)])

print('Model RMSE:', round(model_rmse_sta, 2))
print('Model Mean Absolute Error:', model_mae_sta, 'degrees')
print('Model Accuracy:', round(model_accuracy_sta, 2), '%.')
